More `TweetDispatcher` tests. Needs some reworking
There was a bit of an issue when the mock child factory returned the
testActor that didn't manifest itself until this point in time. The
`TweetDispatcher` code didn't change, the test that failed didn't
change, the common test code didn't change, just the amount of tests
changed. Maybe I was abusing testActor, maybe the tests aren't as
isolated as I hope.